+++ src/tools/maketiledexr/main.cpp
@@ -41,8 +41,8 @@
 #include "foundation/image/pixel.h"
 #include "foundation/image/progressiveexrimagefilewriter.h"
 #include "foundation/image/tile.h"
-#include "foundation/platform/types.h"
 #include "foundation/platform/system.h"
+#include "foundation/platform/types.h"
 #include "foundation/utility/autoreleaseptr.h"
 #include "foundation/utility/log.h"
 
@@ -51,8 +51,8 @@
 #include "OpenEXR/ImfFrameBuffer.h"
 #include "OpenEXR/ImfHeader.h"
 #include "OpenEXR/ImfPixelType.h"
-#include "OpenEXR/ImfTileDescription.h"
 #include "OpenEXR/ImfThreading.h"
+#include "OpenEXR/ImfTileDescription.h"
 
 // Standard headers.
 #include <cstddef>
@@ -76,8 +76,9 @@ int main(int argc, const char* argv[])
     SuperLogger logger;
     Application::check_installation(logger);
 
-    // OpenEXR threading
-    setGlobalThreadCount(System::get_logical_cpu_core_count());
+    // OpenEXR threading.
+    setGlobalThreadCount(
+        static_cast<int>(System::get_logical_cpu_core_count()));
 
     CommandLineHandler cl;
     cl.parse(argc, argv, logger);
